i219: accurate count of library databases and staff
Thanks to @maxkadel for finding the issue -- since we were limiting
our query to only 3 results, the count was only ever 0-3.

This approach has a tradeoff: we are now reading more records into
memory each time.  It will likely never be a huge number of records
(in the low hundreds at the very most), but an alternative approach
could be to issue two db queries: one for the count, and a second
for the first 3 records. Or maybe there is some Rails magic that I
don't know about.