[LibraryDatabases] Handle decomposed diacritics in more_link
The rails transliterate method handles precomposed diacritics (when a
single character is used to represent a letter and its diacritic) just
fine, but when it finds decomposed diacritics (when a separate character
is used for the character and another for its diacritic), it converts
the diacritic into a question mark.

Instead, this converts all diacritics to the decomposed version, then
deletes the diacritics.