Replace rswag with our own OpenAPI handling
This implementation has a lot of similarities with RSwag's API,
although methods are renamed with an openapi_ prefix to distinguish
them from methods with the same name in Rack::Test, rspec-rails,
and Rails.

This implementation makes more assumptions and is less flexible
than RSwag, but hopefully it is a good enough start that we can
extend as needed.

As with RSwag, you can run the tests as normal with RSpec.  If
you want to refresh the OpenAPI/Swagger spec, you can run RSpec
with a specific environment variable (GENERATE_OPEN_API_SPECS=true),
or, more conveniently, `bundle exec rake openapi:generate`

Closes #405