Get smoke tests passing again
While #448 addressed the immediate problem of being able to turn on/off
the banner, it also broke the smoke tests. Requiring flipper-sequel
checks for an active database connection, but the container that runs
the smoke tests shouldn't need to be connected to a database just to
check the health of allsearch-api-staging.

This PR corrects the situation in two ways, either of which should
be sufficient:
* Setting `flipper_sequel` back to `require: false` in the Gemfile,
  and instead adding the needed `require` to a Rails initializer
* Removing the `require spec_helper` instruction in .rspec, which
  is how flipper was being loaded into the smoke tests in the
  first place.