Update rake_migrations tasks to use the db namespace instead of the rom. Clear the default rake db:migrate task from rails if it's defined so that we can run the rom db:migrate task. This will get only the new migrations from the rom_migrate directory and not the rails migrations that still exist in the old_rails_migrate directory We can safely remove the rails migrations after reviewing the application.
In the rom rake db:migrate task create the schema_migrtations file if
it does not exist.

The Sequel migrator extension by default uses filename.
Use a custom migrate task that manually extracts
the integer version of the files and not the full filename.
Evaluates the file to get the Sequel""Migation object
Applies the migration direction in this case :up
Inserts only the integer version into the table
Be specific that the column is a version and the type is a bigint

Remove the migrations_applied task. It is part of the main custome db:migate task

Clear the rake db:migrate task from the deploy rom_migrate task

Update the servers:start task to be explicit in
migrating the test database

related to [#403]